Now df from dbs equal to df from dat files.
DataFrames obtained from a merged SQL database
weren't equal to the ones obtained from .dat
files. This difference was not only column
order or the presence of the cluture_id column.
Differences:
1. The df generated from .dat files had a column
named 'seed' instead of 'rng_seed' as the ones
generated from .db files. This is now corrected.
2. There was a difference in the number of total
and active stem cells reported by the two Data
Frames. This was due to a bad SQL query, and is
now corrected.