p_infty_processing update.
1. Updated the function plot_p_infty_vs_ps. Now
it has options to: plot with error bars, do the
fitting, take into account the y error
(active_stem_cells_indicator_std), print the
fitted parameters, and save the image.

2. Made a new make_df_fitted_pc_vs_t function
that outputs a DataFrame with the fitted
parameters (pc and c) and their standard
deviations for each time.

3. Updated the plot_fitted_pc_vs_t function to
plot p_c(t) with or without error bars, with the
possibility of saving the image.